adicionados listagem de estatísticas

+++ backend/src/main/java/br/com/sistemaveiculos/dtos/DecadaFabricacaoDTO.java
@@ -1,15 +1,17 @@
 package br.com.sistemaveiculos.dtos;
 
+import java.math.BigInteger;
+
 import lombok.Getter;
 
 @Getter
 public class DecadaFabricacaoDTO {
 
-	private Integer qtdVeiculo;
+	private BigInteger qtdVeiculo;
 	private Integer decada;
 	
 	public DecadaFabricacaoDTO(Object[] obj) {
-		this.qtdVeiculo = (Integer) obj[0];
+		this.qtdVeiculo = (BigInteger) obj[0];
 		this.decada = (Integer) obj[1];
 	}
